requirements.txt
matplotlib==3.10.7
numpy==2.3.4
Pillow==12.0.0
torch==2.8.0+cpu
torchvision==0.23.0+cpu
